refactor(front): moved ComponentNode component to be inside board component file because they share styles and are the same feat
