+++ client/src/components/board/ComponentNode.jsx
